% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{equiv}
\alias{equiv}
\title{FUNCTION_TITLE}
\usage{
equiv(table.equiv, input)
}
\arguments{
\item{table.equiv}{PARAM_DESCRIPTION}

\item{input}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
